requirements.txt
asgiref==3.5.0
Django==4.0.4
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-extensions==3.1.5
docopt==0.6.2
flake8==4.0.1
mccabe==0.6.1
Pillow==9.1.0
pipreqs==0.4.11
pycodestyle==2.8.0
pyflakes==2.4.0
six==1.16.0
sqlparse==0.4.2
yarg==0.1.9
